Move ccl scripts into a ccl folder
