+++ scripts/artifacts/FCMQueuedMessageOutlook.py
@@ -22,7 +22,7 @@ SOFTWARE.
 
 import json
 import pathlib
-from scripts.ccl_android_fcm_queued_messages import FcmIterator
+from scripts.ccl.ccl_android_fcm_queued_messages import FcmIterator
 from scripts.artifact_report import ArtifactHtmlReport
 import scripts.ilapfuncs
