+++ scripts/artifacts/FCMQueuedMessagesDump.py
@@ -27,7 +27,7 @@ import base64
 import blackboxprotobuf
 import gzip
 
-from scripts.ccl_android_fcm_queued_messages import FcmIterator
+from scripts.ccl.ccl_android_fcm_queued_messages import FcmIterator
 from scripts.artifact_report import ArtifactHtmlReport
 import scripts.ilapfuncs
