+++ scripts/artifacts/FCMQueuedMessagesInstagram.py
@@ -21,7 +21,7 @@ SOFTWARE.
 """
 import pathlib
 import json
-from scripts.ccl_android_fcm_queued_messages import FcmIterator
+from scripts.ccl.ccl_android_fcm_queued_messages import FcmIterator
 from scripts.artifact_report import ArtifactHtmlReport
 import scripts.ilapfuncs
