+++ scripts/artifacts/FCMQueuedMessagesTwitter.py
@@ -24,7 +24,7 @@ import sys
 import pathlib
 import datetime
 import json
-from scripts.ccl_android_fcm_queued_messages import FcmRecord, FcmIterator
+from scripts.ccl.ccl_android_fcm_queued_messages import FcmRecord, FcmIterator
 from scripts.artifact_report import ArtifactHtmlReport
 import scripts.ilapfuncs
