+++ scripts/artifacts/FCMQueuedMessagesXbox.py
@@ -24,7 +24,7 @@ import datetime
 import json
 import base64
 import pathlib
-from scripts.ccl_android_fcm_queued_messages import FcmIterator
+from scripts.ccl.ccl_android_fcm_queued_messages import FcmIterator
 from scripts.artifact_report import ArtifactHtmlReport
 import scripts.ilapfuncs
