+++ scripts/artifacts/sharedProto.py
@@ -17,7 +17,7 @@ import sqlite3
 import textwrap
 import blackboxprotobuf
 import traceback
-import scripts.ccl_leveldb
+from scripts.ccl import ccl_leveldb
 from datetime import datetime, timedelta
 
 from scripts.artifact_report import ArtifactHtmlReport
@@ -29,7 +29,7 @@ def get_sharedProto(files_found, report_folder, seeker, wrap_text, time_offset):
 
     in_dirs = set(pathlib.Path(x).parent for x in files_found)
     for in_db_dir in in_dirs:
-        leveldb_records = scripts.ccl_leveldb.RawLevelDb(in_db_dir)
+        leveldb_records = ccl_leveldb.RawLevelDb(in_db_dir)
         
         for record in leveldb_records.iterate_records_raw():
             #print(record.seq, record.user_key, record.value)
